requirements.txt
strands-agents
strands-tools
folium